Bug fix(maybe)
The issue with fprintf has been fixed, possibly just for now, it is highly inconsistent and only an issue sometimes, the issue was solved by commenting out part of tilfoej() and adding it back in, no real reason nor solution has been found. Also avira seriously fucks with my file permissions and the ability to run and compile, make sure to turn it off.